Refactor fitting algorithms into separate classes
